requirements.txt
discord.py>=2.3.2
yt-dlp>=2023.12.30
aiohttp>=3.9.1
asyncpg>=0.29.0
sqlalchemy>=2.0.25
alembic>=1.13.1
fastapi>=0.108.0
uvicorn>=0.25.0
redis>=5.0.1
prometheus-client>=0.19.0
cryptography>=41.0.8
python-dotenv>=1.0.0
pydantic>=2.5.2
click>=8.1.7
colorlog>=6.8.0
psutil>=5.9.6
aiosqlite>=0.19.0
Pillow>=10.1.0